+++ service/controller/release/resource/apps/error.go
@@ -1,6 +1,8 @@
 package apps
 
-import "github.com/giantswarm/microerror"
+import (
+	"github.com/giantswarm/microerror"
+)
 
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",